% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_creation_functions.R
\name{build_annual_location_table}
\alias{build_annual_location_table}
\title{Build annual location table}
\usage{
build_annual_location_table(
  snapshot_table,
  change_table,
  year_start,
  year_end,
  verbose = TRUE
)
}
\arguments{
\item{snapshot_table}{A valid 'snapshot' location table describing the location
hierarchy in years where data is available.}

\item{change_table}{A valid 'change' location table describing changes to the location
hierarchy from one year to the next.}

\item{year_start}{First year of the analysis time series}

\item{year_end}{Final year of the analysis time series}

\item{verbose}{\link{bool, default TRUE} Print messages about table construction progress?}
}
\value{

}
\description{
Construct a table describing the location hierarchy for all years in
the time series.
}
\details{
For more information about the 'change', 'annual', and 'stable' location
tables, refer to the package vignettes and the documentation for
\code{\link{validate_location_table}}
}
