% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_creation_functions.R
\name{build_full_location_table}
\alias{build_full_location_table}
\title{Build full location metadata table}
\usage{
build_full_location_table(
  annual_table,
  stable_table,
  year_start,
  year_end,
  verbose = TRUE
)
}
\arguments{
\item{annual_table}{A valid 'annual' location table describing the location
hierarchy for all years in the years between \code{year_start} and \code{year_end}
(inclusive).}

\item{stable_table}{A valid 'stable' location table that identifies unique geographic
units that have stable boundaries across the time period.}

\item{year_start}{First year of the analysis time series}

\item{year_end}{Final year of the analysis time series}

\item{verbose}{\link{bool, default TRUE} Print messages about table construction progress?}
}
\value{

}
\description{

}
\details{
For more information about the 'annual', 'stable', and 'full' location tables,
refer to the package vignettes and the documentation for
\code{\link{validate_location_table}}
}
