% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_creation_functions.R
\name{build_stable_location_table}
\alias{build_stable_location_table}
\title{Build stable location table}
\usage{
build_stable_location_table(
  annual_table,
  change_table,
  year_start,
  year_end,
  verbose = TRUE
)
}
\arguments{
\item{annual_table}{A valid 'annual' location table describing the location
hierarchy for all years in the years between \code{year_start} and \code{year_end}
(inclusive).}

\item{change_table}{A valid 'change' location table describing changes to the location
hierarchy from one year to the next.}

\item{year_start}{First year of the analysis time series}

\item{year_end}{Final year of the analysis time series}

\item{verbose}{\link{bool, default TRUE} Print messages about table construction progress?}
}
\value{

}
\description{
Construct a table that identifies unique geographic units with stable
boundaries throughout the time period.
}
\details{
For more information about the 'change', 'annual', and 'stable' location
tables, refer to the package vignettes and the documentation for
\code{\link{validate_location_table}}
}
