% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_creation_functions.R
\name{create_location_table_template}
\alias{create_location_table_template}
\title{Create a blank template for a location table}
\usage{
create_location_table_template(table_name)
}
\arguments{
\item{table_name}{\link{char} Name of a valid location table. To check valid location table
names, pull from \code{get_valid_location_table_names()}}
}
\value{
A data.table with a single row of NA data entries, formatted with the correct
field names and data types for the specified location table.
}
\description{
Create a blank location hierarchy table with the appropriate fields and
column types. This function can be used to confirm the proper format for the location
hierarchy snapshot and change history tables that need to be filled by the user.
}
\details{
For more information about the location tables, refer to the package vignettes
and the documentation for \code{\link{validate_location_table}}
}
