% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{get_location_table_field_classes}
\alias{get_location_table_field_classes}
\title{Get CRVS location table field classes}
\usage{
get_location_table_field_classes(table_name)
}
\arguments{
\item{table_name}{\link{char} Name of a valid location table}
}
\value{
Named character vector listing expected class for each required column in
a CRVS location table.
}
\description{
Helper function that returns the classes (integer, character, and so on)
expected for each field in a given location table. Note: This function is not exposed
to the package user.
}
