% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{get_location_table_required_fields}
\alias{get_location_table_required_fields}
\title{Get CRVS location table field names}
\usage{
get_location_table_required_fields(table_name)
}
\arguments{
\item{table_name}{\link{char} Name of a valid location table}
}
\value{
Character vector containing all required field names for a given location table
}
\description{
Helper function that returns that names of all required fields for each
type of CRVS location table. Note: this function is not exposed to the package user.
}
