% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{get_valid_change_types}
\alias{get_valid_change_types}
\title{Get valid change types}
\usage{
get_valid_change_types()
}
\value{
\link{char} Vector of valid change types
}
\description{
This function enumerates all the valid location change types
}
