% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io_functions.R
\name{load_location_table}
\alias{load_location_table}
\title{Load location table from CSV}
\usage{
load_location_table(table_name, file, ...)
}
\arguments{
\item{table_name}{\link{char} Name of the CRVS table that is being loaded}

\item{file}{\link{char} Filepath to read from}

\item{...}{Other arguments to data.table::fread()}
}
\value{
Data.table containing the table with properly formatted columns, where known
}
\description{
Load a location table for CRVS mapping from a CSV file
}
\details{
This function ensures that the class of each required column is the correct
type, but does not otherwise validate the table.
}
