% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{message_data_frame}
\alias{message_data_frame}
\title{Message data.frame}
\usage{
message_data_frame(df)
}
\arguments{
\item{df}{Data.frame to print out}
}
\value{
(No return) prints data.table to stderr stream
}
\description{
Print a data.frame to screen using the \code{message()} function
}
