% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_location_annual_change_consistent}
\alias{validate_location_annual_change_consistent}
\title{Validate consistency between location hierarchies and location changes}
\usage{
validate_location_annual_change_consistent(
  annual_table,
  change_table,
  year_start,
  year_end,
  raise_issues_as_errors = TRUE
)
}
\arguments{
\item{annual_table}{An input data.table for the 'annual' location table}

\item{change_table}{An input data.table for the 'change' location table}

\item{year_start}{\link{integer} First year of the analysis time series}

\item{year_end}{\link{integer} Final year of the analysis time series}

\item{raise_issues_as_errors}{\link{logical, default TRUE} Should any cross-compatibility
issues be raised as errors? If TRUE (the default), this function checks for all
issues and then errors if any were found. If FALSE, this function returns a list of
issues and prints a warning message if any were found.}
}
\value{
Returns a list of internal validity issues. If there are no internal validity
issues, returns an empty list.
}
\description{
Validate the cross-table consistency between location hierarchies IN
particular years and location change tables representing differences ACROSS multiple
years.
}
\details{
This function takes the 'annual' and 'change' location tables. These tables
should have already been validated for internal consistency using the functions
\code{validate_location_annual_table_internal()} and
\code{validate_location_change_table_internal()}. This function checks that the two tables
represent consistent changes across the time series.
}
