% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_location_change_table_internal}
\alias{validate_location_change_table_internal}
\title{Validate location change table}
\usage{
validate_location_change_table_internal(input_table, check_years)
}
\arguments{
\item{input_table}{An input data.table for the 'change' location table}

\item{check_years}{\link{integer} Which years should be checked?}
}
\value{
Returns a list of internal validity issues. If there are no internal validity
issues, returns an empty list.
}
\description{
Validate that a location change table is internally consistent for each
queried year.
}
\details{
This function can be used to check the 'change' location table. It performs
the following checks:
\itemize{
\item All required columns are present and have the correct data types and content
\item Each change type is correctly formatted
\item The changes are numbered sequentially, either overall or by year
\item For border changes, 'swap_with' fields are reciprocal (if A<->B then B<->A)
}
}
