% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_location_full_table_internal}
\alias{validate_location_full_table_internal}
\title{Validate full location lookup table}
\usage{
validate_location_full_table_internal(input_table, check_years)
}
\arguments{
\item{input_table}{An input data.table for the 'full' location table, created using
the \link{\code{build_full_location_table}} function.}

\item{check_years}{\link{integer} Which years should be checked?}
}
\value{
Returns a list of internal validity issues. If there are no internal validity
issues, returns an empty list.
}
\description{
Validate that a full location lookup table is internally consistent
}
