% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_location_hierarchy_change_consistent}
\alias{validate_location_hierarchy_change_consistent}
\title{Validate consistency between location hierarchies and location changes}
\usage{
validate_location_hierarchy_change_consistent(
  hierarchy_table,
  change_table,
  max_level,
  check_years
)
}
\arguments{
\item{hierarchy_table}{An input data.table for the 'snapshot' or 'annual' location
table}

\item{change_table}{An input data.table for the 'change' location table}

\item{max_level}{\link{integer} how many levels of administrative subdivisions should be
checked?}

\item{check_years}{\link{integer} Which years should be checked?}
}
\value{
Returns a list of internal validity issues. If there are no internal validity
issues, returns an empty list.
}
\description{
Validate the cross-table consistency between location hierarchies IN
particular years and location change tables representing differences ACROSS multiple
years.
}
\details{
This function takes two tables: a location hierarchy table (either the
'snapshot' or 'annual' table) and a location change table. These tables should have
already been validated for internal consistency using the functions
\code{validate_location_hierarchy_table_internal()} and
\code{validate_location_change_table_internal()}. This function checks that the two tables
are consistent across all specified years.
}
