% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_location_hierarchy_table_internal}
\alias{validate_location_hierarchy_table_internal}
\title{Validate location hierarchy table}
\usage{
validate_location_hierarchy_table_internal(input_table, check_years)
}
\arguments{
\item{input_table}{An input data.table containing a 'snapshot' or 'annual' location
hierarchy table}

\item{check_years}{\link{integer} Which years should be checked?}
}
\value{
Returns a list of internal validity issues. If there are no internal validity
issues, returns an empty list.
}
\description{
Validate that a location hierarchy table is internally consistent for
each queried year.
}
\details{
This function can be used to check the 'snapshot' and 'annual' location
tables. It performs the following checks:
\itemize{
\item All required columns are present and have the correct data types and content
\item All required location levels are present
\item For multi-level tables, each location at the more-detailed levels have valid
parents and each location at the less-detailed levels have valid children
}
}
