% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_location_snapshot_table_internal}
\alias{validate_location_snapshot_table_internal}
\title{Validate 'snapshot' location hierarchy table}
\usage{
validate_location_snapshot_table_internal(input_table, check_years)
}
\arguments{
\item{input_table}{An input data.table containing a 'snapshot' or 'annual' location
hierarchy table}

\item{check_years}{\link{integer} Which years should be checked?}
}
\value{
Returns a list of internal validity issues. If there are no internal validity
issues, returns an empty list.
}
\description{
Wrapper function to validate the location snapshot table, which is a
location hierarchy table that does not need to have all consecutive years available.
}
