% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_location_stable_table_internal}
\alias{validate_location_stable_table_internal}
\title{Validate stable location hierarchy table}
\usage{
validate_location_stable_table_internal(input_table)
}
\arguments{
\item{input_table}{An input data.table for the 'stable' location table, created using
the \link{\code{build_stable_location_table}} function.}
}
\value{
Returns a list of internal validity issues. If there are no internal validity
issues, returns an empty list.
}
\description{
Validate that a stable location hierarchy table is internally consistent
}
