% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_single_location_change}
\alias{validate_single_location_change}
\title{Validate a single location change}
\usage{
validate_single_location_change(change_table)
}
\arguments{
\item{change_table}{A data.table subset of the 'change' location table, providing all
available rows corresponding to a single location change in a single year}
}
\value{
List of issues corresponding to this particular change
}
\description{
Validate a single change within a location change table
}
\details{
This is a helper function for
\code{\link{validate_location_change_table_internal}} and is not exposed to the
package user. This function does NOT check field names or data types, since these
checks are assumed to happen at the table level.
}
