% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_validation_functions.R
\name{validate_year_range}
\alias{validate_year_range}
\title{Validate analysis years}
\usage{
validate_year_range(year_start, year_end)
}
\arguments{
\item{year_start}{First year of the analysis time series}

\item{year_end}{Final year of the analysis time series}
}
\value{
Returns silently; fails informatively if the year range is invalid
}
\description{
Validate the starting and ending years for the analysis period
}
\details{
Note: this is a helper function that is not exposed to the user
}
