% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table_creation_functions.R
\name{create_location_table_template}
\alias{create_location_table_template}
\title{Create a blank template for a location table}
\usage{
create_location_table_template(table_name)
}
\arguments{
\item{table_name}{\link{char} Name of a valid location table. Valid options include
'snapshot', 'annual', 'change', 'stable', and 'full'}
}
\value{
A data.table with a single row of NA data entries, formatted with the correct
field names and data types for the specified location table.
}
\description{
Create a blank location hierarchy table with the appropriate fields and
column types. This function can be used to confirm the proper format for the location
hierarchy snapshot and change history tables that need to be filled by the user.
}
\details{

}
